크기가 작은 부분 문자열 오류 해결

+++ src/main/java/me/programmer/level1/p147355/Main.java
@@ -5,9 +5,9 @@ public class Main {
 class Solution {
 	public int solution(String t, String p) {
 		int answer = 0;
-		int pInt = Integer.parseInt(p);
+		long pInt = Long.parseLong(p);
 		for(int i = 0; i <= t.length() - p.length(); i++) {
-			int tInt = Integer.parseInt(t.substring(i, i+p.length()));
+			long tInt = Long.parseLong(t.substring(i, i+p.length()));
 			if(tInt <= pInt)
 				answer++;
 		}
